wsBcBAABCAAQBQJckqdQCRBK7hj4Ov3rIwAAdHIIAJwxHxModbCMQgAFRdJghk39  7Xl7r4CLVoMkK+upt0dqprvgcuRlYWWyuAHVP5yye8oMDhZNwpSfWhx+0fLXM+Gy  xfcICBeFkFlg++bMOFwcSAr1MMWpMef7SyYcO5tcCJXzM18bzXaeNjOUVIWzWcCb  XO15BXjpJtJF3HdSCWbs4Iqx4Ubz4rLM8YIf9nlpexXC8E5aUPngy6hP5oqILtQQ  K0hLmEqG+58zz1NbNdUQ/Q6l1QpdlvSsQjDMXLAjBCdECj6jZytdWEDKg9tMsRyf  h3nWqID0pd2EfCMgD54+mGiFyqJa3FqoueRZKEZIgcwMuUkis+HVLOoYlmTuTHs=  =sIj3  -----END PGP SIGNATURE-----
Migrate tf.contrib.seq2seq to addons. (#72)

* Initial commit for seq2seq.

Only the V2 implementation is copied from tf.contrib.seq2seq, and
exported.

Current the beam_search_decoder is not broken due to unknown reason from
c kernel, which need more time to debug.

* Fix the unit test and export for beam_search.

Only export the V1 implemenetation of beam_search_decoder.

* Fix unit test failrue and format issue.

* Fix dependency issue for attentionWrapper.

It was depending on the contrib/framework/tensor_util which is now
ported to tf_addon/util/

* Fix unit test for AttentionWrapper.

Previously the output projection dense layer was initialized to be ones, which then cause the data to sampler.sample() to be same within batch, which is why the expected_sample_id mean() to be 0 (it always pick the first index since all the data among the index are the same). Due to the unknown dependency difference between OSS and internal build, sometimes the data will loss some precision, and cause the sample_id to change randomly. This change update the output dense layer to have random weights, which will then lead to much diverse but determined result.

* Remove the "v2" suffix from all the class names.

Since the V1 implemenetation is totally removed, there is no point to
distinguish between v1 and v2.

* Update README.md for generic guildline, and sample code.

* Fix some typo and slight style improvement.

* Fix the unit test due to a rename of the class.

* Reorg all the py and c lib for seq2seq.

The c kernel has been moved to custom_ops and imported from there.
The python lib has been moved out from /python package, and now placed
directly under seq2seq.

* Reformat all the python module with PEP8.

* Update README with Maintainers.

* Documentation cleanup.

* Remove the internal bug link since it is not available externally.

* Update the shape check for AttentionWrapper.

Remove the existing usage of tensor_util, and replace it with control
deps and assert_equal, which should achieve the same behavior for shape
verification.

* Fix lint and format issue.

* More lint fix.

* More and more lint fix.