Fix LayerNormLSTMCell to meet the behavior in original paper.
1. Based on the math in paper, there are 3 norm performed. One for input
kernel, second one for recurrent kernel, and finally for the carry state
before activation. More importantly, the first 2 norm are based on fused
gates, rather than the existing implementation (individual gates
separately).
2. Adding the bias when normalization is on. I think missing the bias
was an oversight for the original implementation, and I have confirmed
this with original author of the contrib code.
3. Update the unit test with the correct value.