Add fallback flag to seq2seq tests (#380)
* Forces attention wrapper test to use the v1-style training loop.

This test uses a Keras anti-pattern in where the layer optionally accepts the output of another layer in the constructor. We are in the process of refactoring the v2 training loop to make it work with tf.function (faster! better! stronger!), but this code will need to be updated as a result. For now, we are falling back to the v1 training loop with an experimental flag; Scott will look for a longer-term solution when he gets through some pressing release issues.