Update SequenceLoss for TensorFlow 2.2 compatibility
For previous versions, we removed the `reduction` attribute to force
tf.keras to call the main `__call__` method. Otherwise, it calls the
`call` method where we have no way of controlling the reduction logic.

In 2.2, tf.keras has a new way of dealing with loss objects (see the
`LossContainer` class). It now always calls `__call__` and requires
the `reduction` attribute to exist.

This new logic is only enabled for V2 execution mode so this commit
also disables the deprecated V1 graph mode for this test.