Make MultiOptimizer serializable
The current implementation of MultiOptimizer puts grads and vars
inside the optimizer_specs variable. Since the grads and vars are
not serializable, any model that uses MultiOptimizer class cannot
be saved or checkpointed. This PR will only extract the necessary
optimizer_specs that are useful for re-instantiating the
MultiOptimizer instance in get_config() method, leading to a
serializable MultiOptimizer implementation.